walls implemented
walls are now represented as lines and are properly rendered when they come into view.
The rendering demands at least one of the edges of the wall to be present. A pretty costy handler that checks if a wall is present(but not the edges) in the view is also available.